refactor(api): Excel 관련 API `document.ts`로 이동 (#118)
